fix zoom issue on middle graph
The cliptoview option erroneously cuts some samples from the
displayed area. This can be seen on very high zoom level.
See bug report: https://github.com/pyqtgraph/pyqtgraph/issues/247

Disable this option for the middle graph as its purpose is to zoom.
The problem also exists on the low window, but since it is less used for
very high zoom levels, keep the cliptoview setting for this one.
Note that this option can always be disabled at runtime thanks to the
pyqtgraph menu (mouse right-click).

Signed-off-by: Sebastien Jan <sjan@baylibre.com>