Add ACME probes power switch control
Some ACME probes include a power switch, allowing the user to remotely
control the power line monitored by the probe.

Add support for detecting this capability and expose a control on the UI
for the user to manually control this power switch.

IIO exposes a power switch related attribute, so we use it for
controlling the power switch.
However, this attribute is exposed for all probes, regardless of the
existence of a switch on the probe (HE10 probes do not have this
switch).
So we use an XMLRPC call to query the probe for the availability of this
feature (recent ACME FS do propose this service). If not available, the
power switch is not exposed to user.
We also get a more appropriate device name through XMLRPC and use it to
populate the default devicename in the UI.

Tested with ACME FS with and without xmlrpc.

Signed-off-by: Sebastien Jan <sjan@baylibre.com>