check Vshunt scaling factor and add --forcevshuntscale
Mainline kernel scaling factor change for Vshunt may cause problems with
tools not aware of this behavioral change (Vshunt used to be reported in
uV, now reported in mV).

Add a check on the Vshunt scaling value and prevent app start if
corresponding to an old setup (uV scaling). This provides a chance for
the user to read the error message and avoid running tests with bad
data.

Also provide a --forcevshuntscale command-line option to override the
default Vshunt scale and start the application even when this error
is detected.

This change allows running pyqtgraph with ACME releases older than b1 by
forcing the scaling factor (option --forcevshuntscale 0.0025)

Signed-off-by: Sebastien Jan <sjan@baylibre.com>